Catch "too many requests" exception in preview service.
